+++ tests/errors/source/duplicate_without_duplicates.rs
@@ -1,0 +1,6 @@
+use duplicate::*;
+#[duplicate_item(
+	name [SomeStruct];
+)]//duplicate_end
+pub struct name(ty);
+//item_end