+++ tests/no_features/expected/only_global_substitutions.expanded.rs
@@ -1,7 +1,4 @@
 use duplicate::*;
-pub struct SomeStruct(i16);
-pub struct SomeStruct2(i32);
-pub struct SomeStruct3(&'static i64);
 pub struct SomeStruct4(u16);
 pub struct SomeStruct5(u32);
 pub struct SomeStruct6(&'static u64);