+++ tests/no_features/from/insert.rs
@@ -1,16 +1,16 @@
 use duplicate::*;
-#[duplicate_item(
+#[substitute_item(
 	name [SomeName]
 )]//duplicate_end
 pub struct name();
 //item_end
-#[duplicate_item(
+#[substitute_item(
 	name	[SomeName2];
 	ty 		[i32]
 )]//duplicate_end
 pub struct name(ty);
 //item_end
-#[duplicate_item(
+#[substitute_item(
 	name	[SomeName3];
 	rf(ty)	[Vec<ty>];
 )]//duplicate_end